build.sh: Support multiple locations of the os-release file
In stateless operating systems the os-release file might not exist
in /etc/ unless a compatability link is in place, or the user has
overriden the distro defaults.

A layering approach should instead be used, to read a set of known
paths in priority order. If the /etc/os-release file is not found,
the fallback location of /usr/lib/os-release will be used instead.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>