build.sh: Add explicit support for Solus
As Solus is in a longstanding transition away from the Pythonic eopkg
package manager (a fork of PiSI) - some workarounds are required.
In a nut shell, eopkg exit codes aren't applied to info/status commands,
so we must use the Pythonic API.

The 32-bit build dependencies are first calculated, and merged with
the 'system.devel' component required for building any package on
Solus. This is similar to 'build-essential' on other distributions.

With this change, we can safely `./build.sh install` on Solus and
have MANGOHUD=1 working with Vulkan applications, as tested on my
own Solus installation with the Serpent demos in Vulkan mode with
bgfx backend.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>