simplify cross arch preloading
This commit does a couple of things. In the mangohud launch script, the library will be found using $LIB with ld, so the correct path is selected depending on the cpu arch. The same is done for mangohud.json (vulkan layer). In addition some now unused files have been deleted.

Signed-off-by: Stephan Lachnit <stephanlachnit@protonmail.com>