Perform updates of gpuInfo struct in more atomic way
The issue is that value might be set to 0, or original miliC (i.e. 35000),
while the other thread is reading it from the memory.

Instead using single writes to the target location,
only with intended end value.

This is a not a full proper fix (like using mutex between overlay and the thread
doing hw info update), but should alleviate issue of logging sometimes
having value 0, or original value of miliC for temperature.