Use alternate method of checking for distro
Use ID/ID_LIKE instead of NAME which will allow less popular forks of
distros to work properly, for example archlabs, whose ID_LIKE will be
arch. And due to the simpler format this should be more
future-compatible.

It appears that some distros don't have ID_LIKE so it also includes ID
in the DISTRO variable.

There are a few downsides to this approach, notably:
 - The OpenSUSE message won't be able to detect whether the user is on
    leap or tumbleweed, so I made it print the command for both.
 - This requires some more globbing since the ID field will be before
    the ID_LIKE field, and the ID_LIKE field can include multiple
    distros.

https://freedesktop.org/software/systemd/man/os-release.html