Add perf counters based metrics for intel platforms
Unlike AMD or Nvidia intel historically only exposes gpu telemetry with
linux perf counters. As with all perf counters on modern linux these
require root or CAP_PERFMON.

This adds a small helper program that can be setuid/setcap'ed with the
appropriate permissions to export the telemetry that mangohud is
interested in. Most of the gpu metrics were informed by the existing
intel_gpu_top tool we are using, however this also exports power metrics
that the intel_gpu_top tool does not support for dgpu configurations and
ensures that we select the appropriate card for all metrics in a
multi-gpu system.

Also this tool exports CPU power metrics from the perf counters. This
allows power monitoring for Intel and zen+ AMD cpus. But this isnt
hooked in this PR since it seems cpu and gpu metrics collection are
separated in the rest of the mangohud.

Mostly this is just a PoC to show what avoiding taking a third-party
dependency on the rarely packaged intel_gpu_top might require. The meson
setup in particular is just something thrown together that builds.