Add setcap helper
This will setcap the perf stats helper appropriately when installed by
root. It wont fail if you try an unprivileged install, but it means the
perf stats helper also wont work.

Also check that we have the CAP_PERFMON capability in the kernel as this
isnt supported on ubuntu 20. On this platform packagers should setuid as
root.