Rework GPUs to allow for multiple
Using sysfs we iterate through available GPUs and determin which one is the
active GPU through fdinfo of the app.
gpu_stats is rewritten to display all available GPUs and their stats or
the current active gpu using the `active_gpu` parameter.
vram is likewise rewritten to display vram for all GPUs or active gpu.
throttling only displays data for the active GPU as we don't expect an
idling GPU to give relevant throttling information.