opengl-shim
Since this library is preloaded, it will be preloaded into any arbitrary
application in the way to the program/game. When one of these applications
share dependencies with mangohud and they use different version of the library
which are incompatible, a crash happens.
To resolve this we create a shim library that only depends on libc.
The shim preloads and hooks opengl swapbuffers and then loads mangohud.
If there is no swapbuffers, there's no opengl context and no reason to load mangohud.
This way we don't preload into arbitrary applications and prevent crashing.