When setting thread names, give a hint that they are for MangoHud
Because MangoHud is a LD_PRELOAD module or a Vulkan layer, its threads
run as part of some host process, typically a game. Putting a common
prefix on the threads means they can be distinguished from threads
created by some other library loaded into the same host process.

For instance, renaming the D-Bus thread from dbus to mangohud-dbus allows
it to be distinguished from some other library also using a thread to
access D-Bus.

Thread names are limited to 15 bytes + '\0', so some of them needed to be
abbreviated appropriately.

Signed-off-by: Simon McVittie <smcv@collabora.com>