vulkan: Expand cases where can_get_driver_info is true
VK_KHR_driver_properties requires VK_KHR_get_physical_device_properties2
extension.  For Vulkan 1.2 both extensions are core, so nothing to do,
for Vulkan 1.1, the latter is core, so only need the former.  For Vulkan
1.0 both extensions are required.

VK_KHR_driver_properties is a device extension that doesn't need to be
explicitly enabled here.  As long as its supported it can be used in the
query on the physical device, before the logical device is created (i.e.
before device extensions can be enabled).  See
https://registry.khronos.org/vulkan/specs/latest/html/vkspec.html#fundamentals-validusage-pNext.

VK_KHR_get_physical_device_properties2 is an instance extension,
required for the previous one.  If the application hasn't enabled it,
do in its behalf.  Also make sure when the extension is used, the KHR
version of the symbol is loaded.

This fixes `vulkan_driver` option in applications that ask for
Vulkan 1.0 (e.g. vkcube) but that the driver does support the extra
extensions.