FEX: Update FEX stats semantics.
FEX updated some [guarantees](https://github.com/FEX-Emu/FEX/pull/4984)
to enforce 16-byte size alignment of the thread stats, and introduced a
new variable for declaring the size of the struct.

This codifies the guarantee that the fex struct will stay 16-byte
aligned sizes, so we get ARM's single-copy atomicity guarantees, and if
FEX decides to add new members to the struct then things won't suddenly
break.

FEX can't guarantee that the API isn't going to break, but it is going
to attempt only adding values to the end as new values get added, so it
is less likely.

A minor change to the `atomic_copy_thread_stats` function as well to
ensure we're taking avantage of that single-copy atomicity.