amdgpu: Unify sysfs readout
[Why]
We do the same thing over and over, just the file handle is different.

[How]
We can use some inline functions which we can call instead. Since we
inline them, the resulting code should be identical.

[Note]
One could think about templating this, but passing the format
specifier as constant to cumbersome. One could also use only one of
the functions, and parse all values as int64, but then we might get
issues with floats.

Signed-off-by: Soeren Grunewald <soeren.grunewald@gmx.net>