Unify reading sysfs files
[Why]
Apparently we have a lot of places in the code where we do basically
the same thing: rewind, fflush, fscanf. So it makes sense to provide
some common code for that.

[How]
Add a new helper function to file_utils.h, which bundles the usual
steps. Since we wanna be able to handle types correctly we use some
templating here, so the compiler will figure the correct format
string based on the requested type.
And since we have C++17 we can use std::optional to make the error
propagation somewhat nicer, allowing for easy fallback values.

Lastly move some includes from the header to the source file, where
the they are actually needed. And while at it also remove the using
namespace directive from the header, since it is considered bad
practice [1] in unscoped headers.

Signed-off-by: Soeren Grunewald <soeren.grunewald@gmx.net>
--
[1] https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rs-using-directive