fileutils: Refactor read_as_int to make it more generic
[Why]
While browsing through the code it became clear that we have
more spots the function could be used, but with different types.
Therefor we switch to a templated approach, to make it easier
to use different format specifiers. This also has the benefit
that we also inline, no matter what.

Signed-off-by: Soeren Grunewald <soeren.grunewald@gmx.net>